+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [1.5.3] - 2023-11-16
 ### Added
--  Added unit tests for BIP39
+- Added unit tests for BIP39
+- Added unit tests for BIP39 word list and SSKR word list
 
 ### Changed
 -